backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9
python-decouple==3.8
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
groq>=0.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.13.0

# WebSocket & Real-Time Messaging
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Production Dependencies
gunicorn>=21.2.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
